move to eth dir. add gitignore
